cad-predictor/requirements.txt
tensorflow==2.0.0
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.0.3
h5py==2.10
requests
paho-mqtt
pillow
scikit-image
protobuf==3.20.*